Java clients, when reading a double from a JsonReader keep it a double

+++ samples/client/petstore/java/okhttp-gson-parcelableModel/src/main/java/org/openapitools/client/model/TypeHolderExample.java
@@ -124,7 +124,7 @@ public class TypeHolderExample implements Parcelable {
 
       @Override
       public NumberItemEnum read(final JsonReader jsonReader) throws IOException {
-        Float value = (Float) jsonReader.nextDouble();
+        Double value = jsonReader.nextDouble();
         return NumberItemEnum.fromValue(String.valueOf(value));
       }
     }
